Game Done -Assets Jump and Slide

+++ app/src/main/java/com/lira/laridaritugas/playgame/ParallaxImage.java
@@ -110,7 +110,7 @@ public class ParallaxImage {
         }
     }
 
-    public void setSpeed(int speed) {
+    public void setSpeed(float speed) {
         this.speed = speed;
     }
 }
